requirements.txt
python-telegram-bot==21.7
environs==14.1.0
google-cloud-dialogflow==2.41.0
